Aborting add/update/delete if pre_add/pre_update/pre_delete raises
The previous behavior would just bubble up exceptions in pre_* BaseView
methods.
The new behavior flashes the error raised, aborts the operation but does
not fail.

Arguably completely silencing any `Exception` may not be the expected
behavior. An alternative would be to only abort on a specific type of
Exception provided by FAB. Something like a
FABAbortPreOperationException. Users would have to raise this or a
derivative of it. I personally thing that this second approach is over
thinking it, the current PR is probably simpler as is.