Fix form actions not working under a prefix for CompactCRUDMixin
The CompactCRUDMixin class handles form actions using the session. The
form action was set to `request.full_path`. This is find as long as
the script_root of the flask app is empty. If it is not, the action
will point to a url that doesn't exist.

Change this to be `request.script_root + request.full_path`. Not using
a more advanced url joiner as script_root always should have no slash
at the end while full_path always will have a slash.