Make updating perms on init optional
Whenever we re-deploy Superset at Airbnb, we restart 5 web servers which
all have a set of many Gunicorn workers.

Each one of these workers is a process and inits the FAB app, resulting
in a fair amount of database contention, and sometimes ultimately
leading to bad database state as I suspect that the perms merge
operations aren't well transaction insulated. We've added unique
constraints on the DB preventing that, but that's not a great solution.

This PR adds a flag to AppBuilder allowing to turn off perm updates on
init.

The default behavior stays unchanged though.