This was a lot harder than I though ;)
Determining relationships is a lot harder than expected. On the "many"
side we need to check the "uselist" attribute of the remote side to
check whether this is a one-to-one relationship. But the remote side
doesn't always have a reference back (no backref or back_populates), in
which case it definitely isn't a one-to-one relationship.

The blanket "except" statements in the functions in the end were what
screwed me over. An error was silently caught, while still resulting
in the test always evaluating to False.