Show widget template: Add some basic blocks
With the help of these blocks it is now possible to override parts
of the show template.

For example it is now possible to create a custom template that extends
the default show.html widget template to add a second set of action
buttons to the top of the show widget area. This makes it easier to
reach the actions in case the page is longer.