Fix unit tests
The tests do not seem to properly tear down test data in the database
after a successful (or failed) test run, causing the next test run to
fail if I don't manually clean up the database in the meantime. The same
issue exists on the master branch, so it is not a consequence of my
changes.