Ignore LDAP search referrals
LDAP referrals are completely broken anyway, and we've already
set ldap.OPT_REFERRALS to 0, so there is no reason to care about
these.

Thus the presence of referrals alone should not cause
authentication of a user to fail due to multiple results being
returned, as they can be filtered out.

This fixes issue #1581.