Handle integrity fails if groups map to same roles
If AUTH_USER_REGISTRATION is enabled and AUTH_ROLE_MAPPING is set such
that multiple LDAP groups map to the same role, then upon attempting to
register the user, specifically inserting that user's roles into the
database, then these inserts will fail due to the contraint check.

To ensure this does not occur, the roles that are to be inserted for the
user needs to be de-dupicated.

Fixes issue #1603.