Fix for DateTimePicker rendering in forms
DateTimePicker renders fine (date plus time) for filters but not for forms (e.g. add) with DateTime fields.
This seems to be mainly due to disabling the ability to pick times ({pickTime: false}) in ab.js but not in ab_filter.js
This fix uniforms ab.js datetimepicker options to the ones (working) in ab_filter.js